backend/requirements.txt
anthropic>=0.35.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.7.0
